% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_metabocano}
\alias{run_metabocano}
\title{Run the Metabocano Application}
\usage{
run_metabocano()
}
\description{
Run the Metabocano Application locally. Shiny deployment
\href{https://plyush1993.shinyapps.io/Metabocano/}{here}.
}
